Adapt reservation service
Adapted the reservation microservice to use the room service interface. Also,
changed a few integer ID fields to long fields, to ensure consistency.

Ref #37

+++ lib/common/src/main/java/nl/tudelft/sem11b/clients/BuildingClient.java
@@ -25,7 +25,7 @@ public class BuildingClient implements BuildingService {
     }
 
     @Override
-    public Optional<BuildingModel> getBuilding(int id) throws ApiException {
+    public Optional<BuildingModel> getBuilding(long id) throws ApiException {
         return api.get("/buildings/" + id, new TypeReference<BuildingModel>() {}).toOptional();
     }
 }
