+++ lib/common/src/main/java/nl/tudelft/sem11b/data/models/BuildingModel.java
@@ -6,7 +6,7 @@ import nl.tudelft.sem11b.data.ApiTime;
  * Holds all building information.
  */
 public class BuildingModel {
-    private final int id;
+    private final long id;
     private final String prefix;
     private final String name;
     private final ApiTime open;
@@ -21,7 +21,7 @@ public class BuildingModel {
      * @param open Time of day at which the building opens
      * @param close This of day at which the building closes
      */
-    public BuildingModel(int id, String prefix, String name, ApiTime open, ApiTime close) {
+    public BuildingModel(long id, String prefix, String name, ApiTime open, ApiTime close) {
         this.id = id;
         this.prefix = prefix;
         this.name = name;
