+++ lib/common/src/main/java/nl/tudelft/sem11b/data/models/EquipmentModel.java
@@ -1,15 +1,15 @@
 package nl.tudelft.sem11b.data.models;
 
 public class EquipmentModel {
-    private final int id;
+    private final long id;
     private final String name;
 
-    public EquipmentModel(int id, String name) {
+    public EquipmentModel(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
